Revert "support translate skill text at battle page (#275)"
This reverts commit 8d38d6159b00bded4d43e5580cc29dbedf9b13d9.
