+++ src/config.js
@@ -17,7 +17,6 @@ const config = {
   fontBold: false,
   transApi: 'caiyun',
   timeout: 8,
-  delay: 300,
   plainText: false,
   autoDownload: false,
   bottomToolbar: false,
