+++ src/modules/content-html.js
@@ -1,7 +1,6 @@
 import getCommHtmlData from '../store/common-html'
 import getArchiveData from '../store/archive'
 import insertSettingHtml from '../setting/insertHtml'
-import { race } from '../utils'
 
 const replaceHTML = async (html, pathname) => {
   let _html = html
@@ -67,7 +66,7 @@ const getHtml = async (encodedHtml, pathname) => {
   return encodeURIComponent(html)
 }
 
-const transHTML = async function (data, pathname) {
+export default async function transHTML(data, pathname) {
   if (data.data) {
     data.data = await getHtml(data.data, pathname)
   }
@@ -84,5 +83,3 @@ const transHTML = async function (data, pathname) {
   }
   return data
 }
-
-export default race(transHTML)
