+++ src/store/skill-job.js
@@ -1,6 +1,5 @@
 import fetchData from '../fetch'
 import parseCsv from '../utils/parseCsv'
-import { getLocalData, setLocalData } from './local-data'
 import filter from '../utils/XSSFilter'
 import { trim } from '../utils/'
 
@@ -9,26 +8,15 @@ let loaded = false
 
 const getSkillData = async (id) => {
   if (!loaded) {
-    let csv = await getLocalData('job-skill')
-    if (!csv) {
-      csv = await fetchData('/blhxfy/data/job-skill.csv')
-      setLocalData('job-skill', csv)
-    }
+    const csv = await fetchData('/blhxfy/data/job-skill.csv')
     const list = parseCsv(csv)
     list.forEach(item => {
       if (item && item.id) {
         const _id = trim(item.id)
-        const _en = trim(item.en)
-        const _ja = trim(item.ja)
-        if (_id) {
-          const value = {
-            name: filter(trim(item.name)),
-            detail: filter(trim(item.detail))
-          }
-          skillMap.set(_id, value)
-          if (_ja) skillMap.set(_ja, value)
-          if (_en) skillMap.set(_en, value)
-        }
+        if (_id) skillMap.set(_id, {
+          name: filter(trim(item.name)),
+          detail: filter(trim(item.detail))
+        })
       }
     })
     loaded = true
