+++ src/store/skill-npc.js
@@ -3,7 +3,6 @@ import parseCsv from '../utils/parseCsv'
 import sortKeywords from '../utils/sortKeywords'
 import filter from '../utils/XSSFilter'
 import { trim } from '../utils/'
-import { getLocalData, setLocalData } from './local-data'
 
 const skillMap = new Map()
 
@@ -23,7 +22,6 @@ const keys = ['skill-1', 'skill-2', 'skill-3', 'skill-4', 'special']
 const state = {
   status: 'init',
   cStatus: 'init',
-  locSkMap: false,
   skillMap,
   skillKeys,
   skillData: null,
@@ -49,36 +47,6 @@ const getCommSkillMap = async () => {
   state.cStatus = 'loaded'
 }
 
-const saveSkillMap = async (skillMap) => {
-  const arr = [...skillMap].slice(-20)
-  setLocalData('skill-npc', JSON.stringify(arr))
-}
-
-const getSkillMap = async () => {
-  const str = await getLocalData('skill-npc')
-  try {
-    const arr = JSON.parse(str)
-    state.skillMap = new Map(arr)
-    state.locSkMap = true
-  } catch (e) {
-
-  }
-}
-
-const saveSkillPath = async (skillData) => {
-  setLocalData('skill-path', JSON.stringify(skillData))
-}
-
-const getSkillPath = async () => {
-  const str = await getLocalData('skill-path')
-  try {
-    const data = JSON.parse(str)
-    state.skillData = data
-  } catch (e) {
-
-  }
-}
-
 const setSkillMap = (list, stable = true) => {
   let npcId, active, idArr
   for (let row of list) {
@@ -100,16 +68,12 @@ const setSkillMap = (list, stable = true) => {
     }
   }
   state.skillMap.set(npcId, skillData)
-  saveSkillMap(state.skillMap)
 }
 
 const getSkillData = async (npcId) => {
-  if (!state.locSkMap) await getSkillMap()
-  if (state.skillMap.has(npcId)) return state
-  await getSkillPath()
+  await getCommSkillMap()
   if (!state.skillData) {
     state.skillData = await fetchData('/blhxfy/data/skill.json')
-    saveSkillPath(state.skillData)
   }
   const csvName = state.skillData[npcId]
   if (csvName) {
@@ -144,4 +108,4 @@ const getLocalSkillData = (npcId) => {
 }
 
 export default getSkillData
-export { skillKeys, getLocalSkillData, getCommSkillMap }
+export { skillKeys, getLocalSkillData }
