+++ src/translate.js
@@ -7,7 +7,6 @@ import transHTML from './modules/content-html'
 import transTownInfo from './modules/town-info'
 import transIslandInfo from './modules/island-info'
 import transChat from './modules/chat-preset'
-import transBattle from './modules/battle'
 import transBuff from './modules/buff'
 import pageIndex, { replaceHour } from './modules/page-index'
 import showVoiceSub from './modules/voice-sub'
@@ -60,16 +59,8 @@ export default async function translate(state) {
       data = await transIslandInfo(data, pathname)
     } else if (pathname.includes('/rest/sound/mypage_voice')) {
       await showVoiceSub(data, pathname, 'list')
-    } else if (/\/rest\/(multi)?raid\/start\.json/.test(pathname)) {
+    } else if (pathname.includes('/rest/multiraid/start.json')) {
       data = await transChat(data)
-      data = await transBattle(data)
-    } else if (
-      /\/rest\/(multi)?raid\/ability_result\.json/.test(pathname)
-      || /\/rest\/(multi)?raid\/temporary_item_result\.json/.test(pathname)
-      || /\/rest\/(multi)?raid\/normal_attack_result\.json/.test(pathname)
-      || /\/rest\/(multi)?raid\/summon_result\.json/.test(pathname)
-    ) {
-      data = await transBattle(data, 'result')
     } else if (/\/rest\/.*?raid\/condition\/\d+\/\d\/\d\.json/.test(pathname)) {
       await transBuff(data.condition)
     } else if (pathname.includes('/user/status')) {
