+++ src/utils/index.js
@@ -96,30 +96,6 @@ const isDomain = (str) => {
   return true
 }
 
-const getPlusStr = (str) => {
-  let plusStr = ''
-  let plusStr2 = ''
-  let _str = str
-  while (_str.endsWith('+') || _str.endsWith('＋')) {
-    plusStr += '＋'
-    plusStr2 += '+'
-    _str = _str.slice(0, _str.length - 1)
-  }
-  return [plusStr, plusStr2, _str]
-}
-
-const race = (func, delay = 300) => {
-  return function (...args) {
-    const promise1 = func(...args)
-    const promise2 = new Promise(rev => {
-      setTimeout(() => {
-        rev(args[0])
-      }, delay)
-    })
-    return Promise.race([promise1, promise2])
-  }
-}
-
 export {
   trim,
   tryDownload,
@@ -129,7 +105,5 @@ export {
   splitSingleLineSkill,
   isDomain,
   removeTag,
-  removeHtmlTag,
-  getPlusStr,
-  race
+  removeHtmlTag
 }
